requirements.txt
astroid==2.5.1
attrs==20.3.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
packaging==20.8
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.1
toml==0.10.2
wrapt==1.12.1
yapf==0.31.0